requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
firebase-admin==6.2.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Talisman==1.1.0
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2
